Expand test suite to be able to test `--no-download`
The test suite can only test success cases, so it needs a bit of adjustement
and a new dependency on temporary in order to setup the environment for
`--no-download` testing.

I've left the existing tests to run in the default data directory, although
this means it's not a hermetic test, I think.