Change to a more modern approach for the footer
The footer positioning still required specifying a fixed size for the
padding, i.e. `4em`, this is not only an annoying magic number to
maintain but also brittle. On small screens for example this might not
be enough if the footer content wraps.

Using the css flex properties remedies this problem. The footer will can
now change height without ever overlapping the results.

Again, I want to thank @entuland (entuland over on #css@freenode) for
coming up with this solution.