Update Hoogle and wiki URLs to HTTPS version in documentation
To eliminate all 301s http://www.haskell.org/hoogle was updated to
https://hoogle.haskell.org. As a result the JSON payload in the API
documentation was updated to show the latest JSON structure which
differed to the existing version.

The URL to jquery in the HTML form embedding examples was 404-ing
so I substituted it with the jquery URL used on the production
Hoogle site.