Serve static files via wai-app-static
After pondering #367 for a while, this seemed like a nicer solution
than providing an option to disable this:

wai-app-static is pretty battletested at this point, it’s well
maintained, has relatively few dependencies and it provides some extra
functionality around caching which may be useful.

That way, we avoid our own validation logic which was broken several
times and I’m not surprised if it’s broken again.

This PR does not address haddock: and file: links. Those deliberately
do not limit paths but they are also disabled by default so from a
security pov this is less of an issue.