Action/Server: Sort any leading `Internal` module to the back
After reversing the module list, we have a pretty good topological
order for modules, meaning the main link of a search result
target (and the first module displayed) will usually be the definition
of that target.

There is however a convention to expose `.Internal` modules so that
users can use unstable APIs if they have a need to.

So in case the first module is such an `.Internal` module, we want to
sort it back and display the next module in the list first.

For example the result

```
toList :: IntSet -> [Key]
containers Data.IntSet.Internal Data.IntSet
```

will now be displayed as

```
toList :: IntSet -> [Key]
containers Data.IntSet Data.IntSet.Internal
```

and link to `Data.IntSet` by default.