Migrate parsing from haskell-src-exts to ghc-lib-parser
Hoogle uses haskell-src-exts both for parsing declarations
and as an internal representation of them (see EDecl constructor of data Entry).
The problem is that haskell-src-exts was not updated since 2020 and fell
behind modern Haskell syntax, uncapable to parse a significant portion of Hackage. ghc-lib-parser is a modern and supported alternative.

Complete migration to ghc-lib-parser would be a mammoth task, so this commit
implements only half of it: use ghc-lib-parser for parsing, but convert
received HsDecl back to Decl from haskell-src-exts, so that the rest
of Hoogle can carry on unchanged.