Fix sizeOf in instance Storable Fingerprint
The commit 2d38e16c9977ce72c5e0c3ac68cb1dea8fe76fcd changed
sizeOf (_ :: Fingerprint) from 64 to 14. This is dangerous,
because it means that pokeByteOff could be reading Word32
with an offset not divisible by sizeOf (_ :: Word32).
Some architectures are forgiving with regards to unaligned reads,
some (such as ARM) are not.

Let's bump sizeOf (_ :: Fingerprint) from 14 to 16.