Fix parseCabalTarball to return the latest version, not the latest revision
parseCabalTarball relies on the assumption that the highest version
of a package is the last in 01-index.tar.gz. This is not necessarily
true in general (one can make a patch release in a long-term support
major version), but especially not true since Hackage gained a concept
of package revisions.

Each package revision is a separate entry in 01-index.tar.gz and it is
very common to revise old releases (because, say, they became incompatible
with new versions of dependencies). Thus Hoogle ends up with a random,
some times terribly outdated version of a package which just happened
to be revised recently.