Fix parsing of build-depends field in Cabal files
hoogle-*.warn file contains 1000+ warnings caused by non-compliant parsing
of build-depends field in Cabal files. E. g., it warns that

hinstaller.cabal: Import of non-existant package Cabal>=1.2.3.0

The actual problem is that Hoogle expects a package name
to be separated from version bounds by a space, which is not
a requirement imposed by Cabal. So Hoogle literally looks
for a package named "Cabal>=1.2.3.0", not "Cabal".

After this commit the number of warnings decreases to ~100
and they all become genuine. E. g.,

cabal-install.cabal: Import of non-existant package Cabal-QuickCheck

Indeed, Cabal-QuickCheck is not on Hackage.